source file reordering
- non-core library definitions moved to fixed_point_utils.h
- accompanying test file, fixed_point_utils_test.cpp
- proposal-specific tests moved to proposal_test.cpp
- testing-specific definitions in test_utils.h
